+++ src/python/pants/backend/project_info/cloc_test.py
@@ -2,9 +2,9 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 from pants.backend.jvm.target_types import JavaLibrary
+from pants.backend.project_info import cloc
 from pants.backend.python.target_types import PythonLibrary
 from pants.core.util_rules import archive, external_tool
-from pants.project_info import cloc
 from pants.testutil.goal_rule_test_base import GoalRuleResult, GoalRuleTestBase
