+++ src/python/pants/backend/project_info/dependencies.py
