+++ src/python/pants/backend/project_info/filedeps_test.py
@@ -11,13 +11,13 @@ from pants.backend.jvm.targets.java_library import JavaLibrary as JavaLibraryV1
 from pants.backend.jvm.targets.jvm_app import JvmApp as JvmAppV1
 from pants.backend.jvm.targets.jvm_binary import JvmBinary as JvmBinaryV1
 from pants.backend.jvm.targets.scala_library import ScalaLibrary as ScalaLibraryV1
+from pants.backend.project_info import filedeps
 from pants.backend.python.target_types import PythonLibrary
 from pants.backend.python.targets.python_library import PythonLibrary as PythonLibraryV1
 from pants.build_graph.build_file_aliases import BuildFileAliases
 from pants.build_graph.resources import Resources as ResourcesV1
 from pants.build_graph.target import Target as TargetV1
 from pants.core.target_types import GenericTarget, Resources
-from pants.project_info import filedeps
 from pants.testutil.goal_rule_test_base import GoalRuleTestBase
