+++ src/python/pants/backend/project_info/list_roots_test.py
@@ -3,8 +3,8 @@
 
 import json
 
+from pants.backend.project_info import list_roots
 from pants.engine.fs import Digest, PathGlobs, Snapshot
-from pants.project_info import list_roots
 from pants.source.source_root import SourceRoot, SourceRootConfig
 from pants.testutil.engine.util import MockGet, run_rule
 from pants.testutil.goal_rule_test_base import GoalRuleTestBase
