+++ src/python/pants/option/global_options.py
@@ -314,7 +314,7 @@ class GlobalOptions(Subsystem):
             "--backend-packages2",
             advanced=True,
             type=list,
-            default=["pants.pants_info", "pants.project_info"],
+            default=["pants.backend.pants_info", "pants.backend.project_info"],
             help=(
                 "Register v2 rules from these backends. The backend packages must be present on "
                 "the PYTHONPATH, typically because they are in the Pants core dist, in a "
