+++ tests/python/pants_test/backend/project_info/tasks/test_dependencies.py
@@ -3,12 +3,12 @@
 
 from pants.backend.jvm.targets.jar_library import JarLibrary
 from pants.backend.jvm.targets.java_library import JavaLibrary
+from pants.backend.project_info.dependencies import DependencyType
 from pants.backend.project_info.tasks.dependencies import Dependencies
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.backend.python.targets.python_requirement_library import PythonRequirementLibrary
 from pants.build_graph.target import Target
 from pants.java.jar.jar_dependency import JarDependency
-from pants.project_info.dependencies import DependencyType
 from pants.python.python_requirement import PythonRequirement
 from pants.testutil.task_test_base import ConsoleTaskTestBase
